camera: port Camera_RefreshFromTrigger to TRX
