+++ src/tr1/game/camera/common.h
@@ -7,5 +7,4 @@
 void Camera_Initialise(void);
 void Camera_ResetPosition(void);
 void Camera_UpdateCutscene(void);
-void Camera_RefreshFromTrigger(const TRIGGER *trigger);
 void Camera_MoveManual(void);
